Extract TippyContent
